pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-vault"
version = "0.1.0"
description = "Local encrypted docs/photos vault with privacy-first retrieval"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "sqlcipher3-binary>=0.5.4",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.6"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = "-q"
markers = [
  "live_llm: opt-in smoke tests that perform bounded live local endpoint calls",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
